+++ docstore-api/src/main/java/fr/simplex_software/docstore/api/ProductResource.java
@@ -6,6 +6,6 @@ import io.quarkus.mongodb.rest.data.panache.*;
 
 import java.math.*;
 
-public interface ProductResource extends PanacheMongoRepositoryResource<ProductRepository, Product, BigInteger>
+public interface ProductResource extends PanacheMongoRepositoryResource<ProductRepository, Product, Long>
 {
 }
