+++ docstore-repository/src/main/java/fr/simplex_software/docstore/repository/CustomerRepository.java
@@ -7,6 +7,6 @@ import jakarta.enterprise.context.*;
 import java.math.*;
 
 @ApplicationScoped
-public class CustomerRepository implements PanacheMongoRepositoryBase<Customer, BigInteger>
+public class CustomerRepository implements PanacheMongoRepositoryBase<Customer, Long>
 {
 }
