+++ docstore-repository/src/main/java/fr/simplex_software/docstore/repository/OrderItemRepository.java
@@ -7,6 +7,6 @@ import jakarta.enterprise.context.*;
 import java.math.*;
 
 @ApplicationScoped
-public class OrderItemRepository implements PanacheMongoRepositoryBase<OrderItem, BigInteger>
+public class OrderItemRepository implements PanacheMongoRepositoryBase<OrderItem, Long>
 {
 }
