+++ docstore-service/src/main/java/fr/simplex_software/docstore/service/ProductService.java
@@ -8,12 +8,12 @@ import java.util.*;
 public interface ProductService
 {
   List<Product> findAllProducts();
-  Optional<Product> findProductById (BigInteger id);
+  Optional<Product> findProductById (Long id);
   List<Product> findProductsByName (String name);
   List<Product> findProductsByNameAndPrice (String name, BigDecimal price);
   List<Product> findProductsByAttribute (String key, String value);
   List<Product> findProductsByAttributes (HashMap<String, String> attributes);
   void createProduct (Product product);
-  void updateProduct (BigInteger id, Product product);
+  void updateProduct (Long id, Product product);
   void removeProduct (Product product);
 }
